refactor(timeline): 优化目录结构
